@@ -20,60 +20,45 @@ namespace Alioth.Framework
     [DebuggerDisplay("Type={objectType}")]
     internal class ObjectBuilder : IObjectBuilder, IAliothServiceProvider
     {
-        private IAliothServiceContainer container;
-        private Type objectType;
-        private IDictionary<String, String> parameters;
-        private IDictionary<String, String> properties;
+        private IAliothServiceContainer _container;
+        private Type _objectType;
 
         /// <summary>
         /// Gets or sets the service object class type.
         /// </summary>
         public Type ObjectType
         {
-            get { return objectType; }
+            get => _objectType;
             set
             {
-                #region precondition
                 if (value == null)
                 {
-                    throw new ArgumentNullException("value");
+                    throw new ArgumentNullException(nameof(value));
                 }
-#if NET40
-                if (!value.IsClass)
-#else
+
                 if (!value.GetTypeInfo().IsClass)
-#endif
                 {
-                    throw new ArgumentOutOfRangeException("value", "The specified object type should be a concrete class.");
+                    throw new ArgumentOutOfRangeException(nameof(value), "The specified object type should be a concrete class.");
                 }
-                #endregion
-                this.objectType = value;
+                _objectType = value;
             }
         }
 
         /// <summary>
         /// Gets the parameters dictioinary of the service object.
         /// </summary>
-        public IDictionary<String, String> Parameters
-        {
-            get { return this.parameters; }
-        }
+        public IDictionary<string, string> Parameters { get; } = new Dictionary<string, string>();
 
         /// <summary>
         /// Gets the properties dictioinary of the service object.
         /// </summary>
-        public IDictionary<String, String> Properties
-        {
-            get { return this.properties; }
-        }
+        public IDictionary<string, string> Properties { get; } = new Dictionary<string, string>();
 
         /// <summary>
         /// Initializes a new instance of the class <c>Alioth.Framework.ObjectBuilder</c>.
         /// </summary>
         public ObjectBuilder()
         {
-            this.parameters = new Dictionary<String, String>();
-            this.properties = new Dictionary<String, String>();
         }
 
         /// <summary>
@@ -82,44 +67,38 @@ namespace Alioth.Framework
         /// <param name="objectType">The service object class type.</param>
         public ObjectBuilder(Type objectType) : this()
         {
-            #region precondition
+
             if (objectType == null)
             {
-                throw new ArgumentNullException("value");
+                throw new ArgumentNullException(nameof(objectType));
             }
-#if NET40
-            if (!objectType.IsClass)
-#else
-                if (!objectType.GetTypeInfo().IsClass)
-#endif
+
+            if (!objectType.GetTypeInfo().IsClass)
             {
-                throw new ArgumentOutOfRangeException("value", "The specified object type should be a concrete class.");
+                throw new ArgumentOutOfRangeException(nameof(objectType), "The specified object type should be a concrete class.");
             }
-            #endregion
-            this.objectType = objectType;
+
+            _objectType = objectType;
         }
 
         /// <summary>
         /// Builds a new instance of the service object.
         /// </summary>
         /// <returns>An object instance of the service object class.</returns>
-        public virtual Object Build()
+        public virtual object Build()
         {
-#if NET40
-            ConstructorInfo[] ctors = objectType.GetConstructors(BindingFlags.Public | BindingFlags.NonPublic).ToArray();
-#else
-            ConstructorInfo[] ctors = objectType.GetTypeInfo().DeclaredConstructors.ToArray();
-#endif
+
+            ConstructorInfo[] ctors = _objectType.GetTypeInfo().DeclaredConstructors.ToArray();
             if (ctors.Length > 1)
             {
-                var ctors2 = ctors.Where(o => o.GetCustomAttributes(false).Any(p => p.GetType() == typeof(DepedencyAtrribute))).ToArray();
+                var ctors2 = ctors.Where(o => o.GetCustomAttributes(false).Any(p => p.GetType() == typeof(DepedencyAttribute))).ToArray();
                 if (ctors2.Length > 1)
                 {
-                    throw new InvalidOperationException(String.Format("Too many constructors annotated with Alioth.Framework.DepedencyAtrribute. Type: {0}", ObjectType.AssemblyQualifiedName));
+                    throw new InvalidOperationException($"Too many constructors annotated with Alioth.Framework.DepedencyAtrribute. Type: {ObjectType.AssemblyQualifiedName}.");
                 }
                 else if (ctors2.Length == 0)
                 {
-                    throw new InvalidOperationException(String.Format("Too many constructors but no one is annotated with Alioth.Framework.DepedencyAtrribute. Type: {0}", ObjectType.AssemblyQualifiedName));
+                    throw new InvalidOperationException($"Too many constructors but no one is annotated with Alioth.Framework.DepedencyAtrribute. Type: {ObjectType.AssemblyQualifiedName}");
                 }
                 return Create(ctors2[0]);
             }
@@ -131,36 +110,30 @@ namespace Alioth.Framework
 
         public void Connect(IAliothServiceContainer container)
         {
-            #region precondition
-            if (container == null)
-            {
-                throw new ArgumentNullException("container");
-            }
-            #endregion
-            this.container = container;
+            _container = container ?? throw new ArgumentNullException(nameof(container));
         }
 
         public object GetService(Type serviceType)
         {
-            return container.GetService(serviceType);
+            return _container.GetService(serviceType);
         }
 
         public object GetService(Type serviceType, string name, string version)
         {
-            return container.GetService(serviceType, name, version);
+            return _container.GetService(serviceType, name, version);
         }
 
-        private Object Create(ConstructorInfo ctor)
+        private object Create(ConstructorInfo ctor)
         {
             ParameterInfo[] ps = ctor.GetParameters();
-            Object[] args = new Object[ps.Length];
+            var args = new object[ps.Length];
             for (int i = 0; i < ps.Length; i++)
             {
                 ParameterInfo pi = ps[i];
                 object value = GetParameterValue(pi);
                 args[i] = value;
             }
-            Object instance = ctor.Invoke(args);
+            var instance = ctor.Invoke(args);
             ConnectContainer(instance);
             InjectDepedencyProperties(instance);
             InjectProperties(instance);
@@ -169,14 +142,13 @@ namespace Alioth.Framework
 
         private object GetParameterValue(ParameterInfo pi)
         {
-            Object value = null;
-            DepedencyAtrribute[] depAttrs = (DepedencyAtrribute[])pi.GetCustomAttributes(typeof(DepedencyAtrribute), false);
+            var depAttrs = (DepedencyAttribute[])pi.GetCustomAttributes(typeof(DepedencyAttribute), false);
+            object value;
             if (depAttrs.Length == 0)
             {
                 Type type = pi.ParameterType;
-                String name = pi.Name;
-                String s;
-                if (parameters.TryGetValue(name, out s))
+                var name = pi.Name;
+                if (Parameters.TryGetValue(name, out var s))
                 {
                     value = ParseValue(type, s);
                 }
@@ -186,13 +158,13 @@ namespace Alioth.Framework
                 }
                 else
                 {
-                    throw new ArgumentException(pi.Name, String.Format("Parameter \"{0}\" must be provided. Type: {1}", pi.Name, ObjectType.AssemblyQualifiedName));
+                    throw new ArgumentException($"Parameter \"{pi.Name}\" must be provided. Type: {ObjectType.AssemblyQualifiedName}", pi.Name);
                 }
             }
             else
             {
-                DepedencyAtrribute depAttr = depAttrs[0];
-                value = this.GetService(depAttr.ServiceType, depAttr.ServiceName, depAttr.ServiceVersion);
+                DepedencyAttribute depAttr = depAttrs[0];
+                value = GetService(depAttr.ServiceType, depAttr.ServiceName, depAttr.ServiceVersion);
             }
 
             return value;
@@ -200,35 +172,27 @@ namespace Alioth.Framework
 
         private void ConnectContainer(object instance)
         {
-            IAliothServiceContainerConnector conn = instance as IAliothServiceContainerConnector;
-            if (conn != null)
+            if (instance is IAliothServiceContainerConnector conn)
             {
-                conn.Connect(this.container);
+                conn.Connect(_container);
             }
         }
 
         private void InjectDepedencyProperties(object instance)
         {
-#if NET40
-            PropertyInfo[] properties = objectType.GetProperties(BindingFlags.NonPublic | BindingFlags.Public | BindingFlags.Instance)
-#else
-            PropertyInfo[] properties = objectType.GetTypeInfo()
+            PropertyInfo[] properties = _objectType.GetTypeInfo()
                 .DeclaredProperties
-#endif
-                .Where(p => p.GetCustomAttributes(false).Any(s => s.GetType() == typeof(DepedencyAtrribute)))
+                .Where(p => p.GetCustomAttributes(false).Any(s => s.GetType() == typeof(DepedencyAttribute)))
                 .ToArray();
             foreach (PropertyInfo p in properties)
             {
-#if NET40
-                DepedencyAtrribute attr = (DepedencyAtrribute)p.GetCustomAttributes(typeof(DepedencyAtrribute), false).First();
-#else
-                DepedencyAtrribute attr = p.GetCustomAttributes<DepedencyAtrribute>().First();
-#endif
-                var v = this.GetService(attr.ServiceType, attr.ServiceName, attr.ServiceVersion);
+                DepedencyAttribute attr = p.GetCustomAttributes<DepedencyAttribute>().First();
+                var v = GetService(attr.ServiceType, attr.ServiceName, attr.ServiceVersion);
                 if (v == null)
                 {
                     throw new KeyNotFoundException(
-                        String.Format(
+                        string.Format(
+                            CultureInfo.InvariantCulture,
                             "The sepecified depedency service with a key '{0}' could not be found. Type: {1}",
                             ServiceKey.Create(attr.ServiceType, attr.ServiceName, attr.ServiceVersion), ObjectType.AssemblyQualifiedName));
                 }
@@ -238,46 +202,33 @@ namespace Alioth.Framework
 
         private void InjectProperties(object instance)
         {
-            if (this.Properties.Count > 0)
+            if (Properties.Count > 0)
             {
-#if NET40
-                var t = instance.GetType();
-                foreach (var item in this.Properties)
-                {
-                    PropertyInfo pi = t.GetProperty(item.Key);
-                    if (pi == null)
-                    {
-                        throw new KeyNotFoundException(String.Format("Invalid [Properites] Configuration: could not found a property with the specified name '{0}'. Type: {1}", item.Key, ObjectType.AssemblyQualifiedName));
-                    }
-                    pi.SetValue(instance, ParseValue(pi.PropertyType, item.Value), null);
-                }
-#else
                 TypeInfo t = instance.GetType().GetTypeInfo();
-                foreach (var item in this.Properties)
+                foreach (var item in Properties)
                 {
                     PropertyInfo pi = t.GetDeclaredProperty(item.Key);
                     if (pi == null)
                     {
-                        throw new KeyNotFoundException(String.Format("Invalid [Properites] Configuration: could not found a property with the specified name '{0}'. Type: {1}", item.Key, ObjectType.AssemblyQualifiedName));
+                        throw new KeyNotFoundException(string.Format(CultureInfo.InvariantCulture, "Invalid [Properites] Configuration: could not found a property with the specified name '{0}'. Type: {1}", item.Key, ObjectType.AssemblyQualifiedName));
                     }
                     pi.SetValue(instance, ParseValue(pi.PropertyType, item.Value), null);
                 }
-#endif
             }
         }
 
-        private static Object ParseValue(Type type, string rawString)
+        private static object ParseValue(Type type, string rawString)
         {
-            Object value = null;
-            if (typeof(Boolean).Equals(type))
+            object value = null;
+            if (typeof(bool).Equals(type))
             {
-                value = Boolean.Parse(rawString);
+                value = bool.Parse(rawString);
             }
-            else if (typeof(Byte).Equals(type))
+            else if (typeof(byte).Equals(type))
             {
-                value = Byte.Parse(rawString, CultureInfo.InvariantCulture);
+                value = byte.Parse(rawString, CultureInfo.InvariantCulture);
             }
-            else if (typeof(Char).Equals(type))
+            else if (typeof(char).Equals(type))
             {
                 value = rawString.FirstOrDefault();
             }
@@ -285,43 +236,43 @@ namespace Alioth.Framework
             {
                 value = DateTime.Parse(rawString, CultureInfo.InvariantCulture);
             }
-            else if (typeof(Decimal).Equals(type))
+            else if (typeof(decimal).Equals(type))
             {
-                value = Decimal.Parse(rawString, CultureInfo.InvariantCulture);
+                value = decimal.Parse(rawString, CultureInfo.InvariantCulture);
             }
-            else if (typeof(Double).Equals(type))
+            else if (typeof(double).Equals(type))
             {
-                value = Double.Parse(rawString, CultureInfo.InvariantCulture);
+                value = double.Parse(rawString, CultureInfo.InvariantCulture);
             }
-            else if (typeof(Int16).Equals(type))
+            else if (typeof(short).Equals(type))
             {
-                value = Int16.Parse(rawString, CultureInfo.InvariantCulture);
+                value = short.Parse(rawString, CultureInfo.InvariantCulture);
             }
-            else if (typeof(Int32).Equals(type))
+            else if (typeof(int).Equals(type))
             {
-                value = Int32.Parse(rawString, CultureInfo.InvariantCulture);
+                value = int.Parse(rawString, CultureInfo.InvariantCulture);
             }
-            else if (typeof(Int64).Equals(type))
+            else if (typeof(long).Equals(type))
             {
-                value = Int64.Parse(rawString, CultureInfo.InvariantCulture);
+                value = long.Parse(rawString, CultureInfo.InvariantCulture);
             }
-            else if (typeof(UInt16).Equals(type))
+            else if (typeof(ushort).Equals(type))
             {
-                value = UInt16.Parse(rawString, CultureInfo.InvariantCulture);
+                value = ushort.Parse(rawString, CultureInfo.InvariantCulture);
             }
-            else if (typeof(UInt32).Equals(type))
+            else if (typeof(uint).Equals(type))
             {
-                value = UInt32.Parse(rawString, CultureInfo.InvariantCulture);
+                value = uint.Parse(rawString, CultureInfo.InvariantCulture);
             }
-            else if (typeof(UInt64).Equals(type))
+            else if (typeof(ulong).Equals(type))
             {
-                value = UInt64.Parse(rawString, CultureInfo.InvariantCulture);
+                value = ulong.Parse(rawString, CultureInfo.InvariantCulture);
             }
-            else if (typeof(Single).Equals(type))
+            else if (typeof(float).Equals(type))
             {
-                value = Single.Parse(rawString, CultureInfo.InvariantCulture);
+                value = float.Parse(rawString, CultureInfo.InvariantCulture);
             }
-            else if (typeof(String).Equals(type))
+            else if (typeof(string).Equals(type))
             {
                 value = rawString;
             }
@@ -329,4 +280,3 @@ namespace Alioth.Framework
         }
     }
 }
-
