@@ -13,32 +13,29 @@ namespace Alioth.Framework
     /// Specifies the type of the service object class.
     /// </summary>
     [AttributeUsage(AttributeTargets.Class, AllowMultiple = true, Inherited = false)]
-    public class ServiceTypeAtrribute : Attribute
+    public class ServiceTypeAttribute : Attribute
     {
-        private Type serviceType;
-        private ReferenceType referenceType;
-
         /// <summary>
         /// Gets the service type.
-        /// <see cref="Alioth.Framework.ServiceKey.Type"/>.
+        /// <see cref="ServiceKey.Type"/>.
         /// </summary>
-        public Type ServiceType { get { return this.serviceType; } }
+        public Type ServiceType { get; }
 
         /// <summary>
         /// Gets the service reference type.
-        /// <see cref="Alioth.Framework.ReferenceType"/>.
+        /// <see cref="ReferenceType"/>.
         /// </summary>
-        public ReferenceType ReferenceType { get { return this.referenceType; } }
+        public ReferenceType ReferenceType { get; }
 
         /// <summary>
         /// Initializes a new instance of the class <c>Alioth.Framework.ServiceTypeAtrribute</c>.
         /// </summary>
         /// <param name="serviceType">the service type.</param>
         /// <param name="referenceType">the service reference type.</param>
-        public ServiceTypeAtrribute(Type serviceType, ReferenceType referenceType = ReferenceType.Strong)
+        public ServiceTypeAttribute(Type serviceType, ReferenceType referenceType = ReferenceType.Strong)
         {
-            this.serviceType = serviceType;
-            this.referenceType = referenceType;
+            ServiceType = serviceType;
+            ReferenceType = referenceType;
         }
     }
 }
